Rename: 캘린더, 차트 컴포넌트 각각 모달안으로 이동
