+++ FE/src/components/CalendarModal/CalendarCourousel/Calendar/style.ts
