+++ FE/src/components/CalendarModal/CalendarCourousel/style.ts
