+++ FE/src/components/PriceChartModal/PriceChart/Chart/util.ts
