+++ FE/src/components/PriceChartModal/PriceChart/MultiRangerSlider/index.tsx
