+++ FE/src/components/PriceChartModal/PriceChart/MultiRangerSlider/style.ts
