+++ FE/src/components/PriceChartModal/PriceChart/PriceInfo/index.tsx
