+++ FE/src/components/PriceChartModal/PriceChart/constants.js
