+++ FE/src/components/PriceChartModal/PriceChart/style.ts
